refactor(internalPropUtils): passing to ts
