+++ lib-ts/index.ts
@@ -1,1 +1,2 @@
+export * from './utils/internalPropUtils';
 export * from './utils/tmpColumnUtils';
