+++ lib-ts/utils/internalPropUtils.ts
@@ -1,0 +1,7 @@
+const INTERNAL_PROP_PREFIX = '$';
+
+function isInternalProp(propName: string) {
+  return propName[0] === INTERNAL_PROP_PREFIX;
+}
+
+export { isInternalProp };
