+++ lib/model/modelClone.js
@@ -3,7 +3,7 @@
 const { isObject, cloneDeep } = require('../utils/objectUtils');
 const { hiddenProps } = require('./modelUtils');
 const { defineNonEnumerableProperty } = require('./modelUtils');
-const { isInternalProp } = require('../utils/internalPropUtils');
+const { isInternalProp } = require('../../lib-ts/dist');
 
 function clone(model, shallow, stripInternal) {
   let clone = null;
