+++ lib/model/modelFilter.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const { isObject } = require('../utils/objectUtils');
-const { isInternalProp } = require('../utils/internalPropUtils');
+const { isInternalProp } = require('../../lib-ts/dist');
 
 function omit(model, args) {
   if (args.length === 1 && isObject(args[0])) {
