+++ lib/model/modelSet.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const { isInternalProp } = require('../utils/internalPropUtils');
+const { isInternalProp } = require('../../lib-ts/dist');
 const { splitQueryProps } = require('./modelQueryProps');
 const { isFunction, isString } = require('../utils/objectUtils');
 const { parseRelationsIntoModelInstances } = require('./modelParseRelations');
