+++ lib/model/modelToJson.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const { isInternalProp } = require('../utils/internalPropUtils');
+const { isInternalProp } = require('../../lib-ts/dist');
 const { mergeQueryProps } = require('./modelQueryProps');
 const { isObject, cloneDeep, isFunction } = require('../utils/objectUtils');
 const EMPTY_ARRAY = [];
