+++ lib/queryBuilder/graph/patch/GraphPatchAction.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const { GraphAction } = require('../GraphAction');
-const { isInternalProp } = require('../../../utils/internalPropUtils');
+const { isInternalProp } = require('../../../../lib-ts/dist');
 const { union, difference, isObject, jsonEquals } = require('../../../utils/objectUtils');
 const promiseUtils = require('../../../utils/promiseUtils');
